Effects.Optparse

Description

Provides the MonadOptparse typeclass.

Since: 0.1

Effect

class Monad m => MonadOptparse m where

Effects for optparse-applicative.

The vast majority of optparse-applicative is not re-exported, as most of the interface is pure. This is merely the functions needed to run a parser.

Since: 0.1

Methods

execParser :: HasCallStack => ParserInfo a -> m a

Run a program description.

Parse command line arguments. Display help text and exit if any parse error occurs.

Since: 0.1

customExecParser :: HasCallStack => ParserPrefs -> ParserInfo a -> m a

Run a program description with custom preferences.

Since: 0.1

handleParseResult :: HasCallStack => ParserResult a -> m a

Handle ParserResult.

Since: 0.1

Misc

type OsPath = OsString

Type representing filenames/pathnames.

This type doesn't add any guarantees over OsString.

osPath :: ReadM OsPath

OsPath Option reader.

Since: 0.1

validOsPath :: ReadM OsPath

OsPath Option reader. This includes validation i.e. fails if the path is considered invalid on the given platform.

Since: 0.1